backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.0
apify-client>=1.6.0
openai>=1.10.0
google-generativeai>=0.3.0
pinecone>=3.0.0
python-multipart>=0.0.6
httpx>=0.24.0

